camera, demo scene, disablable super hot reload

+++ Modules/Client/Graphics/Graphics.Core/ITexture2D.cs
@@ -2,6 +2,6 @@
 
 public interface ITexture2D
 {
-    public float Width { get; }
-    public float Height { get; }
+    public int Width { get; }
+    public int Height { get; }
 }
